Initial commit.
I have some more changes here, so *hopefully* what is in this commit
is just the reproduction for the issue I'm running in to, and none of
my attempting to find a fix.

The issue is that multiple services are publishing equivalent types.
`ComponentInfo` is being published by both Services Alpha and Bravo
as well as being on the Gateway. When the gateway stitches all the
schemas together though, it raises an error:
> The name `ComponentInfo` was already registered by another type.
> (HotChocolate.Types.ObjectType<Gateway.ComponentInfo>)