[WIP] Support for `@mergeable` directive.
This adds a new `Common` project which contains some shared GraphQL
support types, including a `@mergeable` directive that allows
resolving conflicts on equivalent types that are published by multiple
schemas.

This directive is specifically intended for common data types, such as
the `ComponentInfo` here, where each service needs to publish it to be
independent, but they all get stitched together on the gateway, which
also publishes it in the public schema.

Still TODO is actually implementing the "same shape" check.