Implement type shape checks.
With this, I believe this example is actually complete. There's
obviously more that I could do here, like beef up the `GraphQLClient`
class to work a bit more nicely. However, this would just be extra
here since the main reason for this was to figure out why the
automatic type merging in HotChocolate wasn't working in this case,
and to find a way to fix/work around it.

TBH, I'm still not sure *why* the auto-merge isn't handling this case.
My main suspicion based on seeing it merge the types from Alpha and
Bravo with the one in the Gateway separately, is that it's failing
because the `ComponentInfo` type also appears on the gateway's schema.
This `@mergeable` directive does what I want though - allowing for
sharing types across multiple services.

FWIW, I also tried removing the `alphaInfo` and `bravoInfo` fields
from the published schemas, and saw that the merge is no longer needed
since then the `ComponentInfo` is not longer on the public schema; in
other words, *only* the gateway publicizes the type. This is another
possible workaround, but could be a problem if you really do need to
publish one type from multiple services independently.